requirements.txt
psycopg[binary]==3.1.18
python-decouple==3.8
